+++ CHANGELOG.md
@@ -9,6 +9,10 @@ Start of documentation: 2023-02-16 / v1.1.0
 
 - reverted from colorama just_fix_windows_console() to init(), such that stripping ANSI-Codes (e.g. on piping) can be enabled.
 
+### Minor Changes
+
+- -`F`, --`FILES` now also shows the size of the stdin content using -`i`.
+
 
 ## [1.1.3] - 2023-02-26